Backport C++20 countl_zero (#128)
Backport C++20 countl_zero

1) countl_zero can be used in peek_0_bits and is 2X faster.
Note: peek_0_bits is not on a hot path, overall gain is limited.

2) Apply updates for clang-tidy v13 (ships with VS 2022 17.1 Preview 2)

3) Increase upper range of supported CMake to 3.22

+++ fuzztest/main.cpp
@@ -85,7 +85,7 @@ int main(const int argc, const char* const argv[]) // NOLINT(bugprone-exception-
 
     while (__AFL_LOOP(100))
     {
-        std::vector<uint8_t> source(1024 * 1024);
+        std::vector<uint8_t> source(static_cast<size_t>(1024) * 1024);
         const size_t input_length = _read(fd, source.data(), static_cast<unsigned int>(source.size()));
         source.resize(input_length);
